requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
websockets==12.0
python-dotenv==1.0.0
loguru==0.7.2
pytest==8.0.0
httpx==0.27.0
aiortc>=1.3.2
aiohttp>=3.8.1
opencv-python==4.8.1.78
numpy==1.24.3
av==10.0.0
python-multipart>=0.0.5
pytest-asyncio>=0.20.3,<1.0.0

# TUI Dependencies
textual>=0.41.0
sounddevice>=0.4.6
rich>=13.0.0
scipy>=1.11.0
pyaudio>=0.2.11
librosa>=0.10.0 